ai_backend/services/furniture.py
# ai_backend/services/furniture.py - FULL REPLACE

import requests
from bs4 import BeautifulSoup
from typing import List
import logging
from ai_backend.models import FurnitureItem, PriceRange

logger = logging.getLogger(__name__)


def scrape_ethnicraft(furniture_type: str, price_range: PriceRange) -> List[FurnitureItem]:
    """Ethnicraft specific scraper - PLACEHOLDER"""
    # TODO: Implement actual scraping after inspecting site
    logger.warning("Ethnicraft scraper not implemented yet")
    return []


def scrape_kavehome(furniture_type: str, price_range: PriceRange) -> List[FurnitureItem]:
    """Kavehome specific scraper - PLACEHOLDER"""
    logger.warning("Kavehome scraper not implemented yet")
    return []


# Scraper mapping
SCRAPERS = {
    "ethnicraft.com": scrape_ethnicraft,
    "kavehome.com": scrape_kavehome,
    # Add more as you implement them
}


def search_furniture(
    theme: str, 
    room_type: str, 
    furniture_types: List[str], 
    price_range: PriceRange
) -> List[FurnitureItem]:
    """
    Search furniture from theme websites
    
    NOTE: Real scraping implementation needed!
    This returns mock data for now.
    """
    from ai_backend.config import THEMES
    
    websites = THEMES.get(theme.upper(), [])
    all_results = []
    
    for website in websites:
        domain = website.replace("https://", "").replace("http://", "").replace("www.", "").split("/")[0]
        scraper = SCRAPERS.get(domain)
        
        if scraper:
            for furniture_type in furniture_types:
                try:
                    results = scraper(furniture_type, price_range)
                    all_results.extend(results)
                except Exception as e:
                    logger.error(f"Scraping failed for {domain}: {e}")
        else:
            logger.warning(f"No scraper for {domain}")
    
    # For testing: Return mock data if no results
    if not all_results:
        logger.info("No scrapers implemented, returning mock data")
        all_results = _get_mock_furniture(furniture_types, price_range)
    
    # Sort by price
    all_results.sort(key=lambda x: x.price)
    return all_results[:10]


def _get_mock_furniture(furniture_types: List[str], price_range: PriceRange) -> List[FurnitureItem]:
    """Mock furniture data for testing"""
    mock_data = []
    for furn_type in furniture_types:
        mock_data.append(FurnitureItem(
            name=f"Modern {furn_type}",
            link=f"https://example.com/products/{furn_type.lower()}",
            price=(price_range.min + price_range.max) / 2,
            image_url=f"https://via.placeholder.com/300?text={furn_type}",
            dimensions={"width": 48, "depth": 24, "height": 30}
        ))
    return mock_data