requirements.txt
# requirements.txt - FULL REPLACE

# FastAPI & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data Validation
pydantic==2.5.2
pydantic-settings==2.1.0

# AWS
boto3==1.33.2
botocore==1.33.2

# AI & Image Processing
replicate==0.25.1
pillow==10.1.0
requests==2.31.0

# Web Scraping
beautifulsoup4==4.12.2
lxml==4.9.3

# Utilities
python-dotenv==1.0.0
httpx==0.27.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1